feat: removed default values and added a local exec for backups

+++ scripts/backup.sh
@@ -2,36 +2,45 @@
 set -euo pipefail
 
 # CONFIGURATION, MODIFY AS NEEDED
-DB_NAME="dbname"
-DEST_USER="user"
-DEST_HOST="host"
-DEST_DIR="/home/user/backups"
-DEST_PORT="22"
-SSH_KEY="/home/user/.ssh/pg_backup"
-AGE_PUB_KEY="/home/user/backup.agepub"
+DB_NAME=""
+DEST_USER=""
+DEST_HOST=""
+DEST_DIR=""
+DEST_PORT=""
+SSH_KEY=""
+AGE_PUB_KEY=""
+OUTPUT_FILE="" # If defined, backup will be saved locally to this file instead of being sent to remote server
 
 TMP_DIR="/tmp/pg_backups"
 DATE="$(date +%F_%H-%M)"
 BACKUP_FILE="${DB_NAME}_${DATE}.dump"
 ENCRYPTED_BACKUP_FILE="${BACKUP_FILE}.age"
 
-# BACKUP AND ENCRYPT
-mkdir -p "$TMP_DIR"
-pg_dump -F c -Z 9 "$DB_NAME" \
-  | age -R "$AGE_PUB_KEY" \
-  > "$TMP_DIR/$ENCRYPTED_BACKUP_FILE"
+if [ -n "$OUTPUT_FILE" ]; then
+  # Local and unique backup only
+  pg_dump -F c -Z 9 "$DB_NAME" \
+    | age -R "$AGE_PUB_KEY" \
+    > "$OUTPUT_FILE"
+else
+  mkdir -p "$TMP_DIR"
 
-# SEND TO REMOTE SERVER
-scp -i "$SSH_KEY" -P "$DEST_PORT" \
-  "$TMP_DIR/$ENCRYPTED_BACKUP_FILE" \
-  "${DEST_USER}@${DEST_HOST}:${DEST_DIR}/"
+  # BACKUP AND ENCRYPT
+  pg_dump -F c -Z 9 "$DB_NAME" \
+    | age -R "$AGE_PUB_KEY" \
+    > "$TMP_DIR/$ENCRYPTED_BACKUP_FILE"
 
-# CLEANUP OLD BACKUPS
-ssh -i "$SSH_KEY" -p "$DEST_PORT" "${DEST_USER}@${DEST_HOST}" "
-  set -e
-  cd '$DEST_DIR'
-  ls -1t ${DB_NAME}_*.dump.age | tail -n +31 | xargs -r rm --
-"
+  # SEND TO REMOTE SERVER
+  scp -i "$SSH_KEY" -P "$DEST_PORT" \
+    "$TMP_DIR/$ENCRYPTED_BACKUP_FILE" \
+    "${DEST_USER}@${DEST_HOST}:${DEST_DIR}/"
 
-# CLEANUP LOCAL FILE
-rm -f "$TMP_DIR/$ENCRYPTED_BACKUP_FILE"
+  # CLEANUP OLD BACKUPS
+  ssh -i "$SSH_KEY" -p "$DEST_PORT" "${DEST_USER}@${DEST_HOST}" "
+    set -e
+    cd '$DEST_DIR'
+    ls -1t ${DB_NAME}_*.dump.age | tail -n +31 | xargs -r rm --
+  "
+
+  # CLEANUP LOCAL FILE
+  rm -f "$TMP_DIR/$ENCRYPTED_BACKUP_FILE"
+fi
